Gamemode status is now displayed(unfinished)
If either Tag or Manhunt is the selected gamemode, than text will appear in the top-ish left-ish corner of the screen letting the player know 1) which gamemode it is and 2) if the gamemode is active.

Known bug: the text is supposed to be color coded to whether it is running or not, but it only displays as white.

Todo: Position the text in the very top left corner of the screen, auto-detect and adjust that positioning if widescreen is enabled or not.